intelligent_document_ingestion/pyproject.toml
[project]
name = "intelligent-document-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.41.4",
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "locust>=2.42.6",
    "pandas>=2.3.3",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis==5.0.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "streamlit-extras>=0.7.8",
    "uvicorn[standard]>=0.38.0",
]
